Unset temperature attribute when in heat_cool
The temperature attribute is now unset when entering heat_cool mode.

When in heat mode, the temperature and target_temp_low will be kept in sync to allow seamless switching between HVAC modes. Likewise when in cool mode, the temperature and target_temp_high will be kept in sync.

These changes are to fix an issue that was introduced in the updated tile card and more info popup in home assistant 2023.10.